+++ src/test/libraryTest/inspection/autocomplete/autocompletePrimitiveType.ts
@@ -4,86 +4,16 @@
 import { expect } from "chai";
 import "mocha";
 import { Inspection } from "../../../..";
-import { Assert } from "../../../../common";
-import { AutocompleteOption, Position, TriedAutocomplete } from "../../../../inspection";
-import { ActiveNode, ActiveNodeUtils } from "../../../../inspection/activeNode";
-import { Constant, Keyword } from "../../../../language";
-import { IParserState, IParserStateUtils, NodeIdMap, ParseContext, ParseError, ParseOk } from "../../../../parser";
-import { CommonSettings, DefaultSettings, LexSettings, ParseSettings } from "../../../../settings";
+import { AutocompleteOption } from "../../../../inspection";
+import { Constant } from "../../../../language";
+import { DefaultSettings } from "../../../../settings";
 import { TestAssertUtils } from "../../../testUtils";
-
-function assertGetAutocompleteOk<S extends IParserState>(
-    settings: CommonSettings,
-    nodeIdMapCollection: NodeIdMap.Collection,
-    leafNodeIds: ReadonlyArray<number>,
-    position: Position,
-    maybeParseError: ParseError.ParseError<S> | undefined,
-): ReadonlyArray<AutocompleteOption> {
-    const maybeActiveNode: ActiveNode | undefined = ActiveNodeUtils.maybeActiveNode(
-        nodeIdMapCollection,
-        leafNodeIds,
-        position,
-    );
-    if (maybeActiveNode === undefined) {
-        return Keyword.StartOfDocumentKeywords;
-    }
-
-    const triedInspect: TriedAutocomplete = Inspection.tryAutocomplete(
-        settings,
-        nodeIdMapCollection,
-        leafNodeIds,
-        maybeActiveNode,
-        maybeParseError,
-    );
-    Assert.isOk(triedInspect);
-    return triedInspect.value;
-}
-
-function assertGetParseOkAutocompleteOk<S extends IParserState = IParserState>(
-    settings: LexSettings & ParseSettings<IParserState>,
-    text: string,
-    position: Position,
-): ReadonlyArray<AutocompleteOption> {
-    const parseOk: ParseOk<S> = TestAssertUtils.assertGetParseOk(
-        settings,
-        text,
-        IParserStateUtils.stateFactory,
-    ) as ParseOk<S>;
-    return assertGetAutocompleteOk(
-        settings,
-        parseOk.state.contextState.nodeIdMapCollection,
-        parseOk.state.contextState.leafNodeIds,
-        position,
-        undefined,
-    );
-}
-
-function assertGetParseErrAutocompleteOk(
-    settings: LexSettings & ParseSettings<IParserState>,
-    text: string,
-    position: Position,
-): ReadonlyArray<AutocompleteOption> {
-    const parseError: ParseError.ParseError = TestAssertUtils.assertGetParseErr(
-        settings,
-        text,
-        IParserStateUtils.stateFactory,
-    );
-    const contextState: ParseContext.State = parseError.state.contextState;
-
-    return assertGetAutocompleteOk(
-        settings,
-        contextState.nodeIdMapCollection,
-        contextState.leafNodeIds,
-        position,
-        parseError,
-    );
-}
 
 describe(`Inspection - Autocomplete - PrimitiveType`, () => {
     it("type|", () => {
         const [text, position]: [string, Inspection.Position] = TestAssertUtils.assertGetTextWithPosition(`type|`);
         const expected: ReadonlyArray<AutocompleteOption> = [];
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseErrAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseErrAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -94,7 +24,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
     it("type |", () => {
         const [text, position]: [string, Inspection.Position] = TestAssertUtils.assertGetTextWithPosition(`type |`);
         const expected: ReadonlyArray<AutocompleteOption> = Constant.PrimitiveTypeConstantKinds;
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseErrAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseErrAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -107,7 +37,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
             `let x = type|`,
         );
         const expected: ReadonlyArray<AutocompleteOption> = [];
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseErrAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseErrAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -120,7 +50,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
             `let x = type |`,
         );
         const expected: ReadonlyArray<AutocompleteOption> = Constant.PrimitiveTypeConstantKinds;
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseErrAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseErrAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -133,7 +63,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
             `type | number`,
         );
         const expected: ReadonlyArray<AutocompleteOption> = Constant.PrimitiveTypeConstantKinds;
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseOkAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseOkAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -148,7 +78,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
             Constant.PrimitiveTypeConstantKind.Null,
             Constant.PrimitiveTypeConstantKind.Number,
         ];
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseErrAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseErrAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -159,7 +89,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
     it("(x|) => 1", () => {
         const [text, position]: [string, Inspection.Position] = TestAssertUtils.assertGetTextWithPosition(`(x|) => 1`);
         const expected: ReadonlyArray<AutocompleteOption> = [];
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseOkAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseOkAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -172,7 +102,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
             `(x as| number) => 1`,
         );
         const expected: ReadonlyArray<AutocompleteOption> = [];
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseOkAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseOkAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -185,7 +115,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
             `(x as | number) => 1`,
         );
         const expected: ReadonlyArray<AutocompleteOption> = Constant.PrimitiveTypeConstantKinds;
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseOkAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseOkAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -198,7 +128,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
             `(x as| nullable number) => 1`,
         );
         const expected: ReadonlyArray<AutocompleteOption> = [];
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseOkAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseOkAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -211,7 +141,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
             `(x as | nullable number) => 1`,
         );
         const expected: ReadonlyArray<AutocompleteOption> = Constant.PrimitiveTypeConstantKinds;
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseOkAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseOkAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -224,7 +154,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
             `(x as nullable| number) => 1`,
         );
         const expected: ReadonlyArray<AutocompleteOption> = [];
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseOkAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseOkAutocompleteOk(
             DefaultSettings,
             text,
             position,
@@ -237,7 +167,7 @@ describe(`Inspection - Autocomplete - PrimitiveType`, () => {
             `(x as nullable num|ber) => 1`,
         );
         const expected: ReadonlyArray<AutocompleteOption> = Constant.PrimitiveTypeConstantKinds;
-        const actual: ReadonlyArray<AutocompleteOption> = assertGetParseOkAutocompleteOk(
+        const actual: ReadonlyArray<AutocompleteOption> = TestAssertUtils.assertGetParseOkAutocompleteOk(
             DefaultSettings,
             text,
             position,
