+++ src/test/testUtils/assertUtils.ts
@@ -4,8 +4,20 @@
 import { expect } from "chai";
 import "mocha";
 import { Assert, Inspection, Lexer, Task } from "../..";
-import { IParserState, IParserUtils, ParseError, ParseOk, TriedParse } from "../../parser";
-import { LexSettings, ParseSettings } from "../../settings";
+import { AutocompleteOption, Position, TriedAutocomplete } from "../../inspection";
+import { ActiveNode, ActiveNodeUtils } from "../../inspection/activeNode";
+import { Keyword } from "../../language";
+import {
+    IParserState,
+    IParserStateUtils,
+    IParserUtils,
+    NodeIdMap,
+    ParseContext,
+    ParseError,
+    ParseOk,
+    TriedParse,
+} from "../../parser";
+import { CommonSettings, LexSettings, ParseSettings } from "../../settings";
 
 // Only works with single line expressions
 export function assertGetTextWithPosition(text: string): [string, Inspection.Position] {
@@ -74,4 +86,68 @@ function assertGetTriedParse<S extends IParserState = IParserState>(
     const lexerSnapshot: Lexer.LexerSnapshot = triedSnapshot.value;
 
     return IParserUtils.tryRead(stateFactoryFn(settings, lexerSnapshot), settings.parser);
+}
+
+export function assertGetParseOkAutocompleteOk(
+    settings: LexSettings & ParseSettings<IParserState>,
+    text: string,
+    position: Position,
+): ReadonlyArray<AutocompleteOption> {
+    const parseOk: ParseOk = assertGetParseOk(settings, text, IParserStateUtils.stateFactory);
+    const contextState: ParseContext.State = parseOk.state.contextState;
+    return assertGetAutocompleteOk(
+        settings,
+        contextState.nodeIdMapCollection,
+        contextState.leafNodeIds,
+        position,
+        undefined,
+    );
+}
+
+export function assertGetParseErrAutocompleteOk(
+    settings: LexSettings & ParseSettings<IParserState>,
+    text: string,
+    position: Position,
+): ReadonlyArray<AutocompleteOption> {
+    const parseError: ParseError.ParseError = assertGetParseErr(settings, text, IParserStateUtils.stateFactory);
+    const contextState: ParseContext.State = parseError.state.contextState;
+
+    return assertGetAutocompleteOk(
+        settings,
+        contextState.nodeIdMapCollection,
+        contextState.leafNodeIds,
+        position,
+        parseError,
+    );
+}
+
+export function assertGetAutocompleteOk<S extends IParserState>(
+    settings: CommonSettings,
+    nodeIdMapCollection: NodeIdMap.Collection,
+    leafNodeIds: ReadonlyArray<number>,
+    position: Position,
+    maybeParseError: ParseError.ParseError<S> | undefined,
+): ReadonlyArray<AutocompleteOption> {
+    const maybeActiveNode: ActiveNode | undefined = ActiveNodeUtils.maybeActiveNode(
+        nodeIdMapCollection,
+        leafNodeIds,
+        position,
+    );
+    if (maybeActiveNode === undefined) {
+        return Keyword.StartOfDocumentKeywords;
+    }
+
+    const triedInspect: TriedAutocomplete = Inspection.tryAutocomplete(
+        settings,
+        nodeIdMapCollection,
+        leafNodeIds,
+        {
+            scopeById: new Map(),
+            typeById: new Map(),
+        },
+        maybeActiveNode,
+        maybeParseError,
+    );
+    Assert.isOk(triedInspect);
+    return triedInspect.value;
 }
